Move some functions from payload.cc to cache.cc
These functions were only ever used by cache.cc, and were not actually
related to Sorbet's payload--they're more generic functions for
interacting with the disk cache.
